GQL: DatasetPermissions: extract to own file
