+++ src/adapter/graphql/src/queries/datasets/dataset.rs
@@ -278,44 +278,3 @@ pub struct PublicDatasetVisibility {
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-#[derive(SimpleObject, Debug, PartialEq, Eq)]
-pub struct DatasetPermissions {
-    collaboration: DatasetCollaborationPermissions,
-    env_vars: DatasetEnvVarsPermissions,
-    flows: DatasetFlowsPermissions,
-    general: DatasetGeneralPermissions,
-    metadata: DatasetMetadataPermissions,
-}
-
-#[derive(SimpleObject, Debug, PartialEq, Eq)]
-pub struct DatasetCollaborationPermissions {
-    can_view: bool,
-    can_update: bool,
-}
-
-#[derive(SimpleObject, Debug, PartialEq, Eq)]
-pub struct DatasetEnvVarsPermissions {
-    can_view: bool,
-    can_update: bool,
-}
-
-#[derive(SimpleObject, Debug, PartialEq, Eq)]
-pub struct DatasetFlowsPermissions {
-    can_view: bool,
-    can_run: bool,
-}
-
-#[derive(SimpleObject, Debug, PartialEq, Eq)]
-pub struct DatasetGeneralPermissions {
-    can_rename: bool,
-    can_set_visibility: bool,
-    can_delete: bool,
-}
-
-#[derive(SimpleObject, Debug, PartialEq, Eq)]
-pub struct DatasetMetadataPermissions {
-    can_commit: bool,
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
