+++ src/adapter/graphql/src/scalars/mod.rs
@@ -12,6 +12,7 @@ mod account;
 mod data_batch;
 mod data_query;
 mod data_schema;
+mod dataset;
 mod dataset_collaboration;
 mod dataset_endpoints;
 mod dataset_env_var;
@@ -37,6 +38,7 @@ pub(crate) use account::*;
 pub(crate) use data_batch::*;
 pub(crate) use data_query::*;
 pub(crate) use data_schema::*;
+pub(crate) use dataset::*;
 pub(crate) use dataset_collaboration::*;
 pub(crate) use dataset_endpoints::*;
 pub(crate) use dataset_env_var::*;
